+++ libnopegl/src/buffer.h
@@ -22,6 +22,8 @@
 #ifndef BUFFER_H
 #define BUFFER_H
 
+#include <stdlib.h>
+
 struct gpu_ctx;
 
 enum {
@@ -39,14 +41,14 @@ enum {
 
 struct buffer {
     struct gpu_ctx *gpu_ctx;
-    int size;
+    size_t size;
     int usage;
 };
 
 struct buffer *ngli_buffer_create(struct gpu_ctx *gpu_ctx);
-int ngli_buffer_init(struct buffer *s, int size, int usage);
-int ngli_buffer_upload(struct buffer *s, const void *data, int size, int offset);
-int ngli_buffer_map(struct buffer *s, int size, int offset, void **datap);
+int ngli_buffer_init(struct buffer *s, size_t size, int usage);
+int ngli_buffer_upload(struct buffer *s, const void *data, size_t size, size_t offset);
+int ngli_buffer_map(struct buffer *s, size_t size, size_t offset, void **datap);
 void ngli_buffer_unmap(struct buffer *s);
 void ngli_buffer_freep(struct buffer **sp);
